+++ packages/lib-react-components/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+## Fixed
+- roving tabindex for `Tabs`, using the `focusgroup` attribute (#7539).
+
 ## [1.16.0] 2026-05-06
 
 ### Added